applying a patch for > 2GB support!

+++ trunk/qhexview.h
@@ -83,8 +83,8 @@ public:
 	int rowWidth() const;
 
 private:
-	int row_width_;        // amount of 'words' per row
-	int word_width_;       // size of a 'word' in bytes
+        unsigned row_width_;        // amount of 'words' per row
+        unsigned word_width_;       // size of a 'word' in bytes
 	QColor address_color_; // color of the address in display
 	bool show_hex_;        // should we show the hex display?
 	bool show_ascii_;      // should we show the ascii display?
@@ -96,7 +96,7 @@ public:
 
 	void setData(const QSharedPointer<QIODevice>& d);
 	void setAddressOffset(address_t offset);
-	void scrollTo(unsigned int offset);
+        void scrollTo(quint64 offset);
 
 	address_t selectedBytesAddress() const;
 	unsigned int selectedBytesSize() const;
@@ -131,15 +131,15 @@ private:
 
 	unsigned int bytesPerRow() const;
 
-	int dataSize() const;
+        qint64 dataSize() const;
 
-	void drawAsciiDump(QPainter &painter, unsigned int offset, unsigned int row, int size, const QByteArray &row_data) const;
-	void drawHexDump(QPainter &painter, unsigned int offset, unsigned int row, int size, int &word_count, const QByteArray &row_data) const;
-	void drawComments(QPainter &painter, unsigned int offset, unsigned int row, int size) const;
+        void drawAsciiDump(QPainter &painter, quint64 offset, unsigned int row, quint64 size, const QByteArray &row_data) const;
+        void drawHexDump(QPainter &painter, quint64 offset, unsigned int row, quint64 size, int &word_count, const QByteArray &row_data) const;
+        void drawComments(QPainter &painter, quint64 offset, unsigned int row, quint64 size) const;
 
-	void drawAsciiDumpToBuffer(QTextStream &stream, unsigned int offset, int size, const QByteArray &row_data) const;
-	void drawHexDumpToBuffer(QTextStream &stream, unsigned int offset, int size, const QByteArray &row_data) const;
-	void drawCommentsToBuffer(QTextStream &stream, unsigned int offset, int size) const;
+        void drawAsciiDumpToBuffer(QTextStream &stream, quint64 offset, quint64 size, const QByteArray &row_data) const;
+        void drawHexDumpToBuffer(QTextStream &stream, quint64 offset, quint64 size, const QByteArray &row_data) const;
+        void drawCommentsToBuffer(QTextStream &stream, quint64 offset, quint64 size) const;
 
 	QString formatAddress(address_t address);
 	QString format_bytes(const QByteArray &row_data, int index) const;
